requirements.txt
backports.csv==1.0.7
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.0
certifi==2019.6.16
chardet==3.0.4
cheroot==6.5.6
CherryPy==18.1.2
cobble==0.1.3
dwml==0.2
feedparser==5.2.1
future==0.17.1
idna==2.8
inflect==0.2.5
jaraco.functools==2.0
lxml==4.4.0
mammoth==1.3.0
more-itertools==7.2.0
mysqlclient==1.4.4
nltk==3.4.5
numpy==1.17.0
Pattern==3.6
pdfminer.six==20181108
portend==2.5
pycryptodome==3.8.2
PyDocX==0.9.10
PyStarDict==0.8
python-docx==0.8.10
pytz==2019.2
pywin32==224
requests==2.22.0
scipy==1.3.1
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.9.3
tempora==1.14.1
unipath==1.1
urllib3==1.25.3
word-forms==1.0
zc.lockfile==2.0
